feat: MemberService 인터페이스 분리 및 MemberServiceImpl 구현체 생성
- MemberService 인터페이스 분리하여 서비스 레이어의 유연성 향상
- MemberServiceImpl에서 MemberService를 구현하도록 변경
- @RequiredArgsConstructor 추가하여 생성자 주입 간소화
- 기존 MemberService 관련 메서드를 MemberServiceImpl로 이동
- 프로젝트 관련 조회 메서드(getProjectsByMember) 인터페이스 및 구현체에 반영
- 기존 서비스 레이어에서 인터페이스를 참조하도록 리팩토링

#Issue: 445
